requirements.txt
fusionauth-client==1.42.0